build: Correct "hello world" problem
Sometimes the caching in github's docker build process seems to restore
the hello world app that we use to cache the downloaded dependencies
from cargo. This definately should not happen, however it does. This
change will remove the file at the end of the layer so it's impossible
for it to be cached