feat: Add better observability to process manager
this is probably the most important thing that the application does and
currently the user has no visibility (easily) about state changes in
that bit of the app. This will correct that by instrumenting those
functions.

Note: There is a panic in the tracing macro that is there for
compatibility with non-result functions. This has been allowed